+++ frontend/src/components/ui/StationDetailsWindow/reviews/cards/ReplyCard.tsx
@@ -9,7 +9,7 @@ import ButtonNext from '@common/ButtonNext';
 import FlexBox from '@common/FlexBox';
 import Text from '@common/Text';
 
-import ReplyModify from '@ui/StationDetailsWindow/reviews/replies/ReplyModify';
+import ReplyModify from '@ui/StationDetailsWindow/reviews/crud/ReplyModify';
 
 import type { Reply } from '@type';
