+++ frontend/src/components/ui/StationDetailsWindow/reviews/cards/ReviewCardSkeleton.tsx
