+++ frontend/src/components/ui/StationDetailsWindow/reviews/previews/ReviewPreview.tsx
@@ -8,8 +8,8 @@ import ButtonNext from '@common/ButtonNext';
 import FlexBox from '@common/FlexBox';
 import Text from '@common/Text';
 
-import ReviewCard from '@ui/StationDetailsWindow/reviews/ReviewCard';
 import ReviewList from '@ui/StationDetailsWindow/reviews/ReviewList';
+import ReviewCard from '@ui/StationDetailsWindow/reviews/cards/ReviewCard';
 import ReviewPreviewSkeleton from '@ui/StationDetailsWindow/reviews/previews/ReviewPreviewSkeleton';
 import UserRatings from '@ui/StationDetailsWindow/reviews/previews/UserRatings';
